pyproject.toml
[project]
name = "tensor-selenium-tests"
version = "0.1.0"
description = ""
authors = [
    {name = "Ruslan8902",email = "araslanov.r.v@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "selenium (>=4.36.0,<5.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-retry (>=1.7.0,<2.0.0)",
    "allure-pytest (>=2.15.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = "--alluredir=allure-results"